requirements.txt
# UserindoBot
# Copyright (C) 2020  UserindoBot Team, <https://github.com/MoveAngel/UserIndoBot.git>
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.

future
emoji
requests
telethon
sqlalchemy
python-telegram-bot==12.8
psycopg2-binary
pythonping
feedparser
pytz
gtts
bs4
lxml
demjson
wget
bs4
telegraph
pillow
psutil
speedtest-cli>=2.0.2
googletrans
wikipedia
spamwatch
bleach
git+https://github.com/TomyPrs/python-markdown2.git
alphabet_detector
tswift
covid